Feature/api

+++ app/app_monitoring/lib/src/modules/monitoring/model/all_check_model.dart
@@ -1,5 +1,5 @@
 class AllCheckModel {
-  double count;
+  int count;
 
   AllCheckModel({required this.count});
